Update replays route to include sqs and update tests and changelog
