+++ packages/api-client/src/index.ts
@@ -10,4 +10,3 @@ export * as providers from './providers';
 export * as reconciliationReports from './reconciliationReports';
 export * as replays from './replays';
 export * as types from './types';
-export * as replaySqsMessages from './replaySqsMessages';
