+++ packages/api-client/tests/test-replays.js
@@ -34,3 +34,34 @@ test('postKinesisReplays calls the callback with the expected object', async (t)
     callback,
   }));
 });
+
+test('replaySqsMessages calls the callback with the expected object and returns the parsed response', async (t) => {
+  const prefix = 'unitTestStack';
+  const queueName = 'myQueue';
+  const payload = {
+    queueName,
+  };
+
+  const expected = {
+    prefix,
+    payload: {
+      httpMethod: 'POST',
+      resource: '/{proxy+}',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      path: '/replays/sqs',
+      body: JSON.stringify(payload),
+    },
+    expectedStatusCode: 202,
+  };
+  const callback = (configObject) => {
+    t.deepEqual(expected, configObject);
+  };
+
+  await t.notThrowsAsync(replaysApi.replaySqsMessages({
+    prefix,
+    payload,
+    callback,
+  }));
+});
