+++ packages/api/app/routes.js
@@ -28,7 +28,6 @@ const migrationCounts = require('../endpoints/migrationCounts');
 const deadLetterArchive = require('../endpoints/dead-letter-archive');
 const { launchpadProtectedAuth } = require('./launchpadAuth');
 const launchpadSaml = require('../endpoints/launchpadSaml');
-const replaySqsMessages = require('../endpoints/replaySqsMessages');
 
 let token = require('../endpoints/token');
 let { ensureAuthorized } = require('./auth');
@@ -36,9 +35,6 @@ if (process.env.FAKE_AUTH === 'true') {
   token = require('./testAuth'); // eslint-disable-line global-require
   ensureAuthorized = token.ensureAuthorized;
 }
-
-// replay archived S3 messages endpoint
-router.use('/replaySqsMessages', ensureAuthorized, replaySqsMessages);
 
 // dead letters endpoint
 router.use('/deadLetterArchive', ensureAuthorized, deadLetterArchive);
