+++ tf-modules/archive/api.tf
@@ -71,7 +71,7 @@ locals {
       protected_buckets                = join(",", local.protected_buckets)
       provider_kms_key_id              = aws_kms_key.provider_kms_key.key_id
       public_buckets                   = join(",", local.public_buckets)
-      ReplaySQSMessagesLambda          = aws_lambda_function.replay_sqs_messages.arn
+      ReplaySqsMessagesLambda          = aws_lambda_function.replay_sqs_messages.arn
       stackName                        = var.prefix
       system_bucket                    = var.system_bucket
       TOKEN_REDIRECT_ENDPOINT          = local.api_redirect_uri
